+++ src/main/java/forestry/arboriculture/TreeConfig.java
@@ -1,9 +1,0 @@
-package forestry.arboriculture;
-
-public class TreeConfig {
-
-	public static float getSpawnRarity() {
-		return 1.0F;
-	}
-}
-